reservation/requirements.txt
spacy
pdfminer.six
fuzzywuzzy[speedup]
flask
transformers
PyPDF2
torch
numpy
pandas 
scikit-learn
nltk
sentence-transformers
pdfplumber
python-docx
pdf2docx